Feedback from efyx on eclipse-cdt code-analyze session.
Former-commit-id: b9ee7ecdbe98707afe1dfb86169ccae86b024031 [formerly 3ef89b168f0aebc3933d62fdc57d4c8751265b24]
Former-commit-id: 9f591b11f7712efd296e98841f897b6d70de053c

+++ av/NativeAV.cpp
@@ -88,7 +88,7 @@ int NativeAVStreamReader::read(void *opaque, uint8_t *buffer, int size)
         return AVERROR_EXIT;
     }
 
-    size_t copied = 0;
+    int copied = 0;
     int avail = (thiz->streamPacketSize - thiz->streamRead);
 
     // Have data inside buffer
